Added CALL, RET, and RESERVE_LOCAL instructions. A function table has been added to the bytecode file header. The ID is now stored in FunctionOverload instead of FunctionSymbol, meaning each function overload now stores its own unique ID. The BuiltinFuncSymbol class has been replaced with BuiltInFuncOverload, so system function overloads now work, but each overload requires a manual ID. Added primitive function flow analysis, which throws an error if a return statement is missing in a non-void function. Programs must now have a Main function as an entry point; many operations were prohibited outside functions.
The semantic analyzer (particularly its Symbols) needs to be rewritten. A construct like: var x = int32 still works.
Separation of classes and their instances is required.

+++ PettyLang/AST/Expression.cs
@@ -9,9 +9,9 @@ public class IntExpression(int number, Position position) : Expression(position)
     public readonly int Number = number;
 }
 
-public class FloatExpression(double number, Position position) : Expression(position)
+public class FloatExpression(float number, Position position) : Expression(position)
 {
-    public readonly double Number = number;
+    public readonly float Number = number;
 }
 
 public class StringExpression(string str, Position position) : Expression(position)
